# backend/docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        example: Message d'erreur
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      message:
        example: Opération réussie
        type: string
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        description: Validation d'email
        type: string
      full_name:
        description: Nom complet requis
        type: string
      id:
        type: integer
      is_admin:
        description: Est-ce un administrateur ?
        type: boolean
      password_hash:
        description: Hash du mot de passe
        type: string
      phone:
        description: Téléphone requis
        type: string
      updatedAt:
        type: string
    required:
    - email
    - full_name
    - password_hash
    - phone
    type: object
...
  requests.CreateRestaurantRequest:
    properties:
      address:
        description: Adresse complète
        example: route de baduel 11
        type: string
      city:
        description: Ville
        example: cayenne
        type: string
      name:
        description: Nom du restaurant (obligatoire)
        example: petit bateau 1
        type: string
      phone_number:
        description: Numéro de téléphone (optionnel, max 15 caractères)
        example: "+32470542125"
        type: string
      postal_code:
        description: Code postal (limité à 10 caractères pour compatibilité internationale)
        example: "97300"
        type: string
      siren:
        description: SIREN (exactement 9 chiffres, unique)
        example: "123456789"
        type: string
    type: object
  requests.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  requests.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      full_name:
        example: patrick
        type: string
      password:
        example: password123
        minLength: 8
        type: string
      phone:
        example: "+32460232425"
        type: string
    required:
    - email
    - full_name
    - password
    - phone
    type: object
...
  requests.UpdateRestaurantRequest:
    properties:
      address:
        description: Adresse complète
        example: route de baduel 12
        type: string
      city:
        description: Ville
        example: remire
        type: string
      name:
        description: Nom du restaurant (obligatoire)
        example: petit bateau 2
        type: string
      phone_number:
        description: Numéro de téléphone (optionnel, max 15 caractères)
        example: "+32470542125"
        type: string
      postal_code:
        description: Code postal (limité à 10 caractères pour compatibilité internationale)
        example: "97301"
        type: string
      siren:
        description: SIREN (exactement 9 chiffres, unique)
        example: "123456780"
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/users:
    get:
      consumes:
      - application/json
      description: Get all users
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all users
      tags:
      - Admin
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user using email and password
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/requests.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  token:
                    type: string
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Authenticate user
      tags:
      - Users
  /api/auth/signup:
    post:
      consumes:
      - application/json
      description: Create a user with the provided details
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requests.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
      summary: Create a new user
      tags:
      - Users
...
  /api/merchants/restaurants:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Données de la demande
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.CreateRestaurantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
  /api/merchants/restaurants/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: integer
      - description: Données de la demande
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateRestaurantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
  /api/restaurants:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
  /api/restaurants/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID du restaurant
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
...
swagger: "2.0"